+++ packages/ripple-ui-core/src/components/primary-nav/components/mega-menu/mega-menu-back-button.vue
@@ -3,7 +3,7 @@ export default { name: 'RplPrimaryNavMegaMenuBackButton' }
 </script>
 
 <script setup lang="ts">
-import RplIcon from '../icon/icon.vue'
+import RplIcon from '../../../icon/icon.vue'
 
 interface Props {
   label: string
@@ -16,7 +16,7 @@ const props = withDefaults(defineProps<Props>(), {})
   <button
     class="
       rpl-primary-nav__mega-menu-back-button
-      rpl-col-12 rpl-type-label-small rpl-type-weight-bold
+      rpl-u-focusable-inline rpl-col-12 rpl-type-p-small rpl-type-weight-bold
     "
   >
     <RplIcon