+++ packages/ripple-ui-core/src/components/primary-nav/constants.ts
@@ -12,6 +12,12 @@ export interface RplPrimaryNavItem {
   active: boolean
 }
 
+export interface RplPrimaryNavActiveItems {
+  level1?: RplPrimaryNavItem
+  level2?: RplPrimaryNavItem
+  level3?: RplPrimaryNavItem
+}
+
 export const RplPrimaryNavItems = [
   {
     id: '1',
@@ -31,7 +37,25 @@ export const RplPrimaryNavItems = [
           }
         ]
       },
-      { id: '5', text: 'Second level C', url: '#' },
+      {
+        id: '55',
+        text: 'Second level C',
+        url: '#',
+        items: [
+          {
+            id: '22',
+            text: 'Third level A',
+            url: '#',
+            items: [
+              {
+                id: '33',
+                text: 'Fourth level A link with some text that will need to wrap',
+                url: '#'
+              }
+            ]
+          }
+        ]
+      },
       { id: '6', text: 'Second level D', url: '#' },
       { id: '7', text: 'Second level E', url: '#' },
       { id: '8', text: 'Second level F', url: '#' },