+++ packages/ripple-ui-forms/src/inputs/input-utils.ts
@@ -1,11 +1,14 @@
 import { markRaw } from 'vue'
 import RplFormInput from './../components/RplFormInput/RplFormInput.vue'
 import RplFormTextarea from './../components/RplFormTextarea/RplFormTextarea.vue'
-import RplFormCheckbox from './../components/RplFormCheckbox/RplFormCheckbox.vue'
-import RplFormCheckboxGroup from './../components/RplFormCheckbox/RplFormCheckboxGroup.vue'
+import RplFormOption from '../components/RplFormOptions/RplFormOption.vue'
+import RplFormCheckboxGroup from './../components/RplFormOptions/RplFormCheckboxGroup.vue'
+import RplFormRadioGroup from './../components/RplFormOptions/RplFormRadioGroup.vue'
 import RplFormDropdown from './../components/RplFormDropdown/RplFormDropdown.vue'
 import RplFormValidationError from './../components/RplFormValidationError/RplFormValidationError.vue'
 import RplFormHelpText from './../components/RplFormHelpText/RplFormHelpText.vue'
+import RplFormContent from '../components/RplFormContent/RplFormContent.vue'
+import RplFormDivider from '../components/RplFormDivider/RplFormDivider.vue'
 import RplFormActions from '../components/RplFormActions/RplFormActions.vue'
 import {
   outer,
@@ -33,11 +36,14 @@ import {
 export const inputLibrary = {
   RplFormInput: markRaw(RplFormInput),
   RplFormTextarea: markRaw(RplFormTextarea),
-  RplFormCheckbox: markRaw(RplFormCheckbox),
+  RplFormOption: markRaw(RplFormOption),
   RplFormCheckboxGroup: markRaw(RplFormCheckboxGroup),
+  RplFormRadioGroup: markRaw(RplFormRadioGroup),
   RplFormDropdown: markRaw(RplFormDropdown),
   RplFormValidationError: markRaw(RplFormValidationError),
   RplFormHelpText: markRaw(RplFormHelpText),
+  RplFormContent: markRaw(RplFormContent),
+  RplFormDivider: markRaw(RplFormDivider),
   RplFormActions: markRaw(RplFormActions)
 }
 
@@ -106,7 +112,7 @@ export const defaultRplFormInputProps = {
   disabled: '$node.context.disabled',
   placeholder: '$node.context.placeholder',
   options: '$node.context.options',
-  name: '$node.context.name',
+  name: '$node.name',
   className: '$node.context.classes.input',
   validationMeta: '$node.props.validationMeta',
   'aria-describedby': '$fns.getAriaDescribedBy()',