+++ packages/ripple-ui-forms/src/plugin.ts
@@ -8,7 +8,10 @@ import {
   checkboxGroup,
   dropdown,
   number,
-  url
+  url,
+  radioGroup,
+  content,
+  divider
 } from './inputs/index'
 
 const rplFormInputs = (node) => {
@@ -66,8 +69,14 @@ rplFormInputs.library = (node) => {
       return node.define(checkbox)
     case 'RplFormCheckboxGroup':
       return node.define(checkboxGroup)
+    case 'RplFormRadioGroup':
+      return node.define(radioGroup)
     case 'RplFormDropdown':
       return node.define(dropdown)
+    case 'RplFormContent':
+      return node.define(content)
+    case 'RplFormDivider':
+      return node.define(divider)
     case 'RplFormActions':
       return node.define(actions)
   }